% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSender.R
\name{send_file}
\alias{send_file}
\title{Send a hardcoded file to a specified URL}
\usage{
send_file()
}
\value{
Response from the server.
}
\description{
This function retrieves a hardcoded file from the local file system and sends it to a hardcoded URL.
}
